build 0.2

+++ Assets/Game/Scripts/UI/Buttons/SkateButton.cs
@@ -1,46 +1,78 @@
 using LazyFramework;
+using TMPro;
 using UnityEngine;
 using UnityEngine.UI;
 
-public class SkateButton : ButtonBase
+public class SkateButton : MonoBehaviour
 {
-    [SerializeField] private Image skateImage;
+    [Header("Library")]
+    [SerializeField] SkateBoardLibrary library;
+
+    [Header("References")]
+    [SerializeField] Image skateImage;
     [SerializeField] Image noti;
     [SerializeField] Image lockIcon;
+
+    [Header("Equip button")]
+    [SerializeField] TMP_Text useText;
+
+    [SerializeField] Sprite useSprite;
+    [SerializeField] Sprite usedSprite;
+    [SerializeField] Sprite buySprite;
+
+    [SerializeField] Image button;
+    [SerializeField] TMP_Text unlockText;
     int index;
+    private void Start()
+    {
+        Subscribe();
+    }
+    private void OnDestroy()
+    {
+        Unsubscribe();
+    }
     private void OnEnable()
     {
         CheckUnlock();
-
+        CheckUse();
     }
-    public override void OnClick()
+    public void OnButtonCLick()
     {
-        base.OnClick();
-        ShowPopupSkate();
-        NotificationService.RemoveSkateNotification(index);
-    }
-
-    void ShowPopupSkate()
-    {
-        GameServices.ShowPopupSkateBoard(index);
+        if (DataService.IsSkateEquiping(index) == false)
+        {
+            DataService.EquipSkate(index);
+            NotificationService.RemoveSkateNotification(index);
+        }
     }
     public void SetData(SkateBoardData data,int index)
     {
         skateImage.sprite = data.board;
         this.index = index;
+
         OnNotificationChange(null);
         CheckUnlock();
+        CheckUse();
     }
 
     private void Lock()
     {
         lockIcon.gameObject.SetActive(true);
         skateImage.color=Color.black;
+
+        button.gameObject.SetActive(false);
+
+        unlockText.gameObject.SetActive(true);
+        unlockText.text=$"Unlock at level {library.GetUnlockLevel(index)}";
     }
     private void Unlock()
     {
+        button.gameObject.SetActive(true);
+        button.sprite=buySprite;
+
         lockIcon.gameObject.SetActive(false);
         skateImage.color=Color.white;
+
+        unlockText.gameObject.SetActive(false);
     }
 
     /// <summary>
@@ -57,6 +89,21 @@ public class SkateButton : ButtonBase
             Unlock();
         }
     }
+
+    private void CheckUse()
+    {
+        if(DataService.IsSkateEquiping(index) == true)
+        {
+            button.sprite = usedSprite;
+            useText.text="Used";
+        }
+        else
+        {
+            button.sprite = useSprite;
+            useText.text = "Use";
+        }
+
+    }
     private void OnNotificationChange(OnNotificationChange e)
     {
         if(NotificationService.IsSkateNoti(index) == true)
@@ -68,13 +115,20 @@ public class SkateButton : ButtonBase
             noti.gameObject.SetActive(false);
         }
     }
+    private void OnSkateUse(OnSkateUse e)
+    {
+        CheckUnlock();
+        CheckUse();
+    }
 
-    protected override void Subscribe()
+    protected void Subscribe()
     {
         Event<OnNotificationChange>.Subscribe(OnNotificationChange);
+        Event<OnSkateUse>.Subscribe(OnSkateUse);
     }
-    protected override void Unsubscribe()
+    protected void Unsubscribe()
     {
         Event<OnNotificationChange>.Unsubscribe(OnNotificationChange);
+        Event<OnSkateUse>.Unsubscribe(OnSkateUse);
     }
 }
